// apps/dashboard/types/systemNotification.ts
export enum toastSeverity {
  success = "success",
  info = "info",
  warn = "warn",
  error = "error",
  secondary = "secondary",
  contrast = "contrast",
}

export enum toastSummary {
  success = "Permintaan Berhasil!",
  info = "info",
  warn = "warn",
  error = "Permintaan Gagal!",
  secondary = "secondary",
  contrast = "contrast",
}

export enum toastAuth {
  fillEmptyField = "Silahkan isi kolom yang masih kosong",
  correctIncorrectField = "Silahkan perbaiki kolom yang salah",
  nameEmpty = "Kolom Nama Masih Kosong",
  emailEmpty = "E-mail Masih Kosong",
  emailNotValid = "E-mail Tidak Valid",
  phoneEmpty = "Nomor Ponsel Masih Kosong",
  phoneNotValid = "Nomor Ponsel Tidak Falid",
  passwordEmpty = "Password Masih Kosong",
  passwordNotValid = "Password Minimal 8 Karakter",
  passwordNotSame = "Password Tidak Sama",
  otpEmpty = "OTP Masih Kosong",
}